Remover notificacoes

+++ src/main/java/com/missedfaces/server/api/NotificationsController.java
@@ -8,6 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -32,9 +33,12 @@ public class NotificationsController {
   }
 
   @RequestMapping(value = "/deletar/{id}", method = RequestMethod.DELETE)
+  @Transactional
   public ResponseEntity<Void> edit(@PathVariable Long id) {
     Notification notification = notificationRepository.findByIdAndUser(id, accountsService.currentUser());
-    notificationRepository.delete(notification);
+    if (notification != null) {
+      notificationRepository.delete(notification);
+    }
     return new ResponseEntity<>(HttpStatus.NO_CONTENT);
   }
